Rename RoomCsv -> RoomUtil. Move RoomUtil into its own roomutil directory. Create ViewModel for RoomUtilScreen and inject database into it.
