+++ app/src/main/java/com/heyzeusv/androidutilitieslibrary/feature/roomutil/RoomUtilViewModel.kt
@@ -1,0 +1,11 @@
+package com.heyzeusv.androidutilitieslibrary.feature.roomutil
+
+import androidx.lifecycle.ViewModel
+import com.heyzeusv.androidutilitieslibrary.database.Database
+import dagger.hilt.android.lifecycle.HiltViewModel
+import javax.inject.Inject
+
+@HiltViewModel
+class RoomUtilViewModel @Inject constructor(
+    private val database: Database
+) : ViewModel()
